DjangoWebProject1/quickstart/urls.py

from rest_framework import routers, serializers, viewsets
from quickstart import views


router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)